Add update password feature
