+++ src/components/Notification/style.module.css
@@ -1,0 +1,17 @@
+.notification {
+  position: fixed;
+  display: grid;
+  gap: .8em;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  background-color: white;
+  border: 1px solid #ccc;
+  padding: 30px;;
+  border-radius: 5px;
+  z-index: 1001; 
+}
+
+.notification button {
+  margin: 10px;
+}
