+++ src/components/PasswordInput/style.module.css
@@ -4,11 +4,6 @@ input::-webkit-inner-spin-button {
   margin: 0;
 }
 
-input[type=number] {
-  appearance: textfield;
-  -moz-appearance: textfield;
-}
-
 input {
   padding-inline: 0.6em;
   border-radius: var(--br-100);
