+++ src/components/UpdatePasswordModal/index.tsx
@@ -4,11 +4,15 @@ import { PasswordInput } from '../PasswordInput';
 import { useUpdatePassword } from '../../hooks/useUpdatePassword';
 import { Button } from '../Button';
 
-export function UpdatePasswordModal() {
+type UpdatePsswordModalProps = {
+  onClose: () => void;
+}
+export function UpdatePasswordModal( {onClose}: UpdatePsswordModalProps) {
   const {
     password,
     newPassword,
     repeatPassword,
+    successMessage,
     updatePasswordErrorMessage,
     handleFieldChange,
     handleUpdatePasswordClick,
@@ -16,27 +20,33 @@ export function UpdatePasswordModal() {
 
   return (
     <Modal
+      onClose={onClose}
       children={
         <div className={styles.wrapper}>
           <h2 className={styles.h2}>Update your password</h2>
           <PasswordInput
             label={'Password'}
-            value={password}
+            value={password!}
+            placeholder={' '}
             onChange={e => handleFieldChange('password', e)}
           />
           <PasswordInput
             label={'New Password'}
-            value={newPassword}
+            value={newPassword!}
+            placeholder={' '}
             onChange={e => handleFieldChange('new-password', e)}
           />
           <PasswordInput
             label={'Repeat New Password'}
-            value={repeatPassword}
+            value={repeatPassword!}
+            placeholder={' '}
             onChange={e => handleFieldChange('repeat-password', e)}
           />
-          <p className={styles.errorMessage}>
-            {updatePasswordErrorMessage && ''}
-          </p>
+          {updatePasswordErrorMessage ? (
+            <p className={styles.errorMessage}>{updatePasswordErrorMessage}</p>
+          ) : (
+            <p className={styles.successMessage}>{successMessage}</p>
+          )}
           <Button
             label={'Change password'}
             onClick={handleUpdatePasswordClick}
