+++ src/components/UpdatePasswordModal/style.module.css
@@ -1,0 +1,26 @@
+.wrapper {
+  display: grid;
+  gap: .6em;
+  justify-items: center;
+}
+
+.h2 {
+  font-size: var(--fs-400);
+  margin-bottom: 1em;;
+}
+
+.errorMessage {
+  color: red;
+  font-size: var(--fs-200);
+  position: unset;
+  height: 32px;
+  width: 234px;
+}
+
+.successMessage {
+  color: green;
+  font-size: var(--fs-200);
+  position: unset;
+  height: 32px;
+  width: 234px;
+}
