+++ src/hooks/index.ts
@@ -1,2 +1,4 @@
-export * from './useLogin'
-export * from './useDashboard'
+export * from './useLogin';
+export * from './useDashboard';
+export * from './useLogout';
+export * from './useUpdatePassword';
