+++ src/validations/index.ts
@@ -2,4 +2,6 @@ export * from './isEmailValid';
 export * from './isRutValid';
 export * from './isPasswordValid';
 export * from './isPhoneValid';
-export * from './isTokenValid';
+export * from './decodeToken';
+export * from './arePasswordsEqual';
+export * from './isTokenExpired'
